Add shared controls including nonce api fetch
