+++ assets/js/data/collections/index.js
@@ -12,7 +12,7 @@ import * as selectors from './selectors';
 import * as actions from './actions';
 import * as resolvers from './resolvers';
 import reducer from './reducers';
-import { controls } from './controls';
+import { controls } from '../shared-controls';
 
 registerStore( STORE_KEY, {
 	reducer,
